Fix circular deps (#2934)
## Description

I resolved some circular dependencies:
```
1) reanimated2/js-reanimated/JSReanimated.ts > reanimated2/js-reanimated/Mapper.ts
2) reanimated2/js-reanimated/JSReanimated.ts > reanimated2/js-reanimated/MapperRegistry.ts
3) reanimated2/animation/util.ts > reanimated2/animation/delay.ts
4) reanimated2/animation/util.ts > reanimated2/animation/repeat.ts
5) reanimated2/animation/util.ts > reanimated2/animation/sequence.ts
6) reanimated2/animation/styleAnimation.ts > reanimated2/animation/timing.ts > reanimated2/animation/util.ts
7) reanimated2/NativeMethods.ts > reanimated2/hook/useAnimatedRef.ts
```
Because files used each other's TS types.

I added [madge](https://github.com/pahen/madge) tool to detect circular dependencies to CI test

Fixes https://github.com/software-mansion/react-native-reanimated/issues/2924

Inspired by @yiheyang 's PR - https://github.com/software-mansion/react-native-reanimated/pull/2925 👏👏👏
