+++ package.json
@@ -24,7 +24,8 @@
     "type:generate:tsc": "yarn tsc",
     "type:generate:remove.ts": "find ./lib -type f -name \"*.ts\" -and -not -name \"*.d.ts\" -delete",
     "type:generate:remove.tsx": "find ./lib -type f -name \"*.tsx\" -and -not -name \"*.d.ts\" -delete",
-    "prepare": "husky install"
+    "prepare": "husky install",
+    "circular_dependency_check": "yarn madge --extensions js,ts,tsx --circular ./src"
   },
   "main": "lib/Animated.js",
   "module": "lib/Animated",
@@ -114,6 +115,7 @@
     "husky": "^7.0.0",
     "jest": "^26.6.3",
     "lint-staged": "^11.2.0",
+    "madge": "^5.0.1",
     "prettier": "^2.2.1",
     "react": "17.0.2",
     "react-native": "0.67.2",
